# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Afrinza Marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  title: string;
  description?: string;
  price: number;
  category: string;
  location: string;
  imageUrl?: string;
  sellerId: number;
  sellerName: string;
  sellerWhatsapp: string;
  sellerAvatar?: string;
  isSponsored: boolean;
  isPremiumSeller: boolean;
  rating: number;
  reviewCount: number;
  stock?: number;
  createdAt?: string;
}

export type ProductDetail = Product & {
  images?: string[];
  deliveryOptions?: string[];
  paymentMethods?: string[];
};

export interface ProductListResponse {
  products: Product[];
  total: number;
}

export interface CreateProductRequest {
  title: string;
  description?: string;
  price: number;
  category: string;
  location: string;
  imageUrl?: string;
  images?: string[];
  sellerId: number;
  stock?: number;
  deliveryOptions?: string[];
  paymentMethods?: string[];
}

export interface Seller {
  id: number;
  storeName: string;
  ownerName: string;
  description?: string;
  location: string;
  categories?: string[];
  whatsapp: string;
  avatarUrl?: string;
  bannerUrl?: string;
  isPremium: boolean;
  rating: number;
  reviewCount: number;
  productCount: number;
  joinedAt?: string;
}

export type SellerDetail = Seller & {
  products?: Product[];
};

export interface SellerListResponse {
  sellers: Seller[];
  total: number;
}

export interface CreateSellerRequest {
  storeName: string;
  ownerName: string;
  description?: string;
  location: string;
  categories?: string[];
  whatsapp: string;
  avatarUrl?: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  icon?: string;
  productCount: number;
}

export interface CategoryListResponse {
  categories: Category[];
}

export interface Review {
  id: number;
  productId: number;
  buyerName: string;
  rating: number;
  comment?: string;
  createdAt: string;
}

export interface ReviewListResponse {
  reviews: Review[];
  averageRating: number;
  total: number;
}

export interface CreateReviewRequest {
  productId: number;
  buyerName: string;
  rating: number;
  comment?: string;
}

export interface CartItem {
  id: number;
  sessionId: string;
  productId: number;
  product?: Product;
  quantity: number;
}

export interface CartResponse {
  items: CartItem[];
  total: number;
  itemCount: number;
}

export interface AddToCartRequest {
  sessionId: string;
  productId: number;
  quantity: number;
}

export interface Order {
  id: number;
  sessionId: string;
  buyerName: string;
  buyerPhone: string;
  buyerAddress?: string;
  items?: CartItem[];
  total: number;
  paymentMethod: string;
  deliveryMethod: string;
  status: string;
  createdAt?: string;
}

export interface CreateOrderRequest {
  sessionId: string;
  buyerName: string;
  buyerPhone: string;
  buyerAddress?: string;
  paymentMethod: string;
  deliveryMethod: string;
}

export interface MarketplaceStats {
  totalSellers: number;
  totalProducts: number;
  totalLocations: number;
  totalCategories: number;
  featuredSellers: number;
}

export type GetProductsParams = {
  category?: string;
  location?: string;
  minPrice?: number;
  maxPrice?: number;
  search?: string;
  sellerId?: number;
  limit?: number;
  offset?: number;
};

export type GetNearbyProductsParams = {
  location: string;
};

export type GetSellersParams = {
  location?: string;
  category?: string;
};

export type GetReviewsParams = {
  productId: number;
};

export type GetCartParams = {
  sessionId: string;
};