# lib/api-zod/src/generated/types/createProductRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Afrinza Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProductRequest {
  title: string;
  description?: string;
  price: number;
  category: string;
  location: string;
  imageUrl?: string;
  images?: string[];
  sellerId: number;
  stock?: number;
  deliveryOptions?: string[];
  paymentMethods?: string[];
}