# lib/api-zod/src/generated/types/createReviewRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Afrinza Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateReviewRequest {
  productId: number;
  buyerName: string;
  rating: number;
  comment?: string;
}