# lib/api-zod/src/generated/types/marketplaceStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Afrinza Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface MarketplaceStats {
  totalSellers: number;
  totalProducts: number;
  totalLocations: number;
  totalCategories: number;
  featuredSellers: number;
}