# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Afrinza Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { CartItem } from "./cartItem";

export interface Order {
  id: number;
  sessionId: string;
  buyerName: string;
  buyerPhone: string;
  buyerAddress?: string;
  items?: CartItem[];
  total: number;
  paymentMethod: string;
  deliveryMethod: string;
  status: string;
  createdAt?: string;
}