# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Afrinza Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  title: string;
  description?: string;
  price: number;
  category: string;
  location: string;
  imageUrl?: string;
  sellerId: number;
  sellerName: string;
  sellerWhatsapp: string;
  sellerAvatar?: string;
  isSponsored: boolean;
  isPremiumSeller: boolean;
  rating: number;
  reviewCount: number;
  stock?: number;
  createdAt?: string;
}