# lib/api-zod/src/generated/types/reviewListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Afrinza Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { Review } from "./review";

export interface ReviewListResponse {
  reviews: Review[];
  averageRating: number;
  total: number;
}