# lib/api-zod/src/generated/types/seller.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Afrinza Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface Seller {
  id: number;
  storeName: string;
  ownerName: string;
  description?: string;
  location: string;
  categories?: string[];
  whatsapp: string;
  avatarUrl?: string;
  bannerUrl?: string;
  isPremium: boolean;
  rating: number;
  reviewCount: number;
  productCount: number;
  joinedAt?: string;
}